Force C++11

+++ Problem Set 1/reference_calc.cpp
@@ -1,5 +1,6 @@
 // for uchar4 struct
 #include <cuda_runtime.h>
+#include <cmath>
 
 void referenceCalculation(const uchar4* const rgbaImage,
                           unsigned char *const greyImage,
@@ -10,7 +11,7 @@ void referenceCalculation(const uchar4* const rgbaImage,
     for (size_t c = 0; c < numCols; ++c) {
       const int pos = r * numCols + c;
       uchar4 rgba = rgbaImage[pos];
-      float channelSum = (.299f * rgba.x + .587f * rgba.y + .114f * rgba.z);
+      float channelSum = (.299f * (float)rgba.x + .587f * (float)rgba.y + .114f * (float)rgba.z);
       greyImage[pos] = channelSum;
     }
   }
